pyproject.toml
[project]
name = "test-api-framework"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "allure-pytest>=2.13.5",
    "clickhouse-sqlalchemy>=0.3.2",
    "jenkins>=1.0.2",
    "jsonpath>=0.82.2",
    "pandas>=2.2.3",
    "paramiko",
    "pymongo",
    "pymysql",
    "pytest>=8.3.2",
    "pytest-rerunfailures>=3.0.0",
    "pyyaml>=6.0.2",
    "redis>=5.2.1",
    "requests>=2.32.3",
    "sqlalchemy>=2.0.36",
    "xlrd>=2.0.1",
    "xlutils>=2.0.0",
    "dbutils>=3.1.1",
    "pytest-tmreport>=1.4.1"
]

[tool.setuptools]
packages = [
    "base",
    "demo",
    "data",
    "conf",
    "common",
    "report",
    "testcase",
    "mock_server"
]